screach: Move verification into its own module
Pure refactor, getting ready for use in GREASE.
