+++ screach/src/Screach/Diagnostic.hs
@@ -20,6 +20,7 @@ import Prettyprinter.Render.Text qualified as PP
 import Screach.Distance.Diagnostic qualified as Distance
 import Screach.GoalEvaluator.Diagnostic qualified as GoalEvaluator
 import Screach.Run.Diagnostic qualified as Run
+import Screach.Verify.Diagnostic qualified as Verify
 import System.IO (stderr)
 import What4.ProgramLoc qualified as WPL
 import Prelude hiding (log)
@@ -33,6 +34,7 @@ data Diagnostic where
   GreaseDiagnostic :: Grease.Diagnostic -> Diagnostic
   RunDiagnostic :: Run.Diagnostic -> Diagnostic
   DistanceDiagnostic :: Distance.Diagnostic -> Diagnostic
+  VerifyDiagnostic :: Verify.Diagnostic -> Diagnostic
   ScheduledSuccessor :: WPL.ProgramLoc -> WPL.ProgramLoc -> Int -> Diagnostic
   ExecutingFrame :: WPL.ProgramLoc -> String -> Diagnostic
   ResumingFrame :: WPL.ProgramLoc -> Diagnostic
@@ -45,6 +47,7 @@ instance PP.Pretty Diagnostic where
       GreaseDiagnostic diag -> PP.pretty diag
       RunDiagnostic diag -> PP.pretty diag
       DistanceDiagnostic diag -> PP.pretty diag
+      VerifyDiagnostic diag -> PP.pretty diag
       ScheduledSuccessor loc fromLoc dist ->
         PP.hsep ["Scheduled:", PP.viaShow loc, "from", PP.viaShow fromLoc, "at distance", PP.pretty dist]
       ExecutingFrame loc stType -> PP.hsep ["Executing frame:", PP.viaShow loc, PP.pretty stType]
@@ -62,6 +65,7 @@ severity =
     GreaseDiagnostic diag -> Grease.severity diag
     RunDiagnostic diag -> Run.severity diag
     DistanceDiagnostic diag -> Distance.severity diag
+    VerifyDiagnostic diag -> Verify.severity diag
     ScheduledSuccessor{} -> Debug
     ExecutingFrame _ _ -> Debug
     ResumingFrame{} -> Debug
